test: split scrape and maintenance integration flows
